% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_protocol.R
\name{breeding_protocol}
\alias{breeding_protocol}
\title{Breeding protocols}
\usage{
breeding_protocol(
  n_population,
  n_lifetime,
  breeding_rotation_rate = 2,
  breeding_rotation = NULL
)
}
\arguments{
\item{n_population}{A vector setting number of individuals in the active populations (the size of the inactive population is the sum this vector).}

\item{n_lifetime}{The number of iterations of the genetic algorithm.}

\item{breeding_rotation_rate}{The rate at which the breeding rotation schedule is applied in the protocol.}

\item{breeding_rotation}{A vector indicating the rotation of the bulls in the population.}
}
\value{
A list containing the 'breeding_rotation_rate' and 'breeding_rotation'.
}
\description{
\code{breeding_protocol} creates a list of all parameters needed to control the breeding schedule of the \link{cbga} algorithm.
}
