\name{CBGA-package}
\alias{CBGA-package}
\alias{CBGA}
\docType{package}
\title{
\packageTitle{CBGA}
}
\description{
\packageDescription{CBGA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CBGA}
\packageIndices{CBGA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CBGA}

Maintainer: \packageMaintainer{CBGA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
