% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbga_optimiser.R
\name{cbga.control}
\alias{cbga.control}
\title{Controlled Breeding Genetic Algorithm Control}
\usage{
cbga.control(
  n_lifetimes = 100,
  n_population = 100,
  n_chromosomes = NULL,
  pi_mutation = NULL,
  pi_recombination = NULL,
  s_age = 4,
  s_mutation = 1,
  trace = TRUE
)
}
\arguments{
\item{n_lifetimes}{The number of iterations of the genetic algorithm.}

\item{n_population}{The number of individuals in both the active and inactive populations.}

\item{n_chromosomes}{A vector assigning the length of the bitstrings (related to the precision of the optimisation).}

\item{pi_mutation}{The probability of mutation.}

\item{pi_recombination}{The probability of recombination.}

\item{s_age}{A scaling factor for the effect of age on the fitness (> 0).}

\item{s_mutation}{A scaling factor for the effect of the number of mutations on the fitness (> 0).}

\item{trace}{TRUE/FALSE: Should an internal trace of the genetic algorithm be shown?}
}
\value{
A list of parameters used for the \link{cbga}-function.
}
\description{
\code{cbga.control} creates a list of all parameters needed to run the genetic algorithm.
}
