% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbga_optimiser.R
\name{cbga}
\alias{cbga}
\title{Controlled Breeding Genetic Algorithm}
\usage{
cbga(f, lower, upper, bp = "proportional", ...)
}
\arguments{
\item{f}{The function to be optimised by the algorithm.}

\item{lower}{A vector of lower bounds.}

\item{upper}{A vector of upper bounds.}

\item{bp}{A string, or \link{bp}-object, defining the breeding protocol used in the genetic algorithm. See details for more information on allowed pre-sets supplied as strings, and \link{set_bp} for more information on possible breeding protocols.}

\item{...}{Additional control parameters used in the genetic algorithm. See \link{cbga.control} for more information.}
}
\value{
The fittest individual found by the genetic algorithm, arranged in the list containing the following:
}
\description{
\code{cbga} is a genetic algorithm used to fit box constrained univariate functions. The genetic algorithm utilises controlled breeding for a more efficient estimation.
}
\details{
The following breeding protocols...
}
\examples{
\dontrun{
f <- function(x) exp(-x^2)
lower <- -1
upper <- 1

cbga(f, lower, upper)
}

}
