% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbga_optimiser.R
\name{cbga}
\alias{cbga}
\title{Controlled Breeding Genetic Algorithm}
\usage{
cbga(f, lower, upper, bp_type = "proportional", ga_pars = list())
}
\arguments{
\item{f}{The function to be optimised.}

\item{lower}{A vector of lower bounds.}

\item{upper}{A vector of upper bounds.}

\item{bp_type}{The breeding protocol used in the genetic algorithm. See details for more information.}

\item{ga_pars}{A list of additional parameters used in the genetic algorithm. See \link{cbga.control} for more information.}
}
\value{
The fittest individual found by the genetic algorithm.
}
\description{
\code{cbga} is a genetic algorithm used to fit box contrained univarite functions. The genetic algorithm utilises controlled breeding for a more efficient estimation.
}
\details{
The following breeding protocols...
}
